+++ templates/parts/navbar_page.php
@@ -1,6 +1,6 @@
 <nav class="navbar navbar-expand-lg navbar-light bg-light">
 	<div class="container justify-content-between">
-		<a class="navbar-brand" href="index.php">Blog</a>
+		<a class="navbar-brand" href="/">Blog</a>
 		<div>
 			<button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false"
 				aria-label="Toggle navigation">
